// src/models/merchant.model.ts
export class MerchantModel {
    id?: number;
    companyName?: string;
    registrationNo?: string;
    addrLine1?: string;
    addrLine2?: string;
    postcode?: string;
    city?: string;
    state?: string;
    country?: string;
    tel?: string;
    fax?: string;
    email?: string;
    bankName?: string;
    bankAccNo?: string;
}